Add inventory and money for user-service

+++ src/main/kotlin/com/esop/schema/Order.kt
@@ -1,37 +1,37 @@
 package com.esop.schema
 
 class OrderFiller{
-    var quantity: Long = 0
-    var amount: Long = 0
-    constructor(quantity: Long, amount: Long){
+    var quantity: Int = 0
+    var amount: Int = 0
+    constructor(quantity: Int, amount: Int){
         this.quantity = quantity
         this.amount = amount
     }
 }
 
 class Order{
-    var quantity: Long = 0
+    var quantity: Int = 0
     var type: String = "NULL" // BUY or SELL
-    var price: Long = 0
+    var price: Int = 0
     var orderId: Int = 0
     var timeStamp = System.currentTimeMillis()
 
     var orderStatus: String = "PENDING" // COMPLETED, PARTIAL, PENDING
-    var currentQuantity: Long = 0
+    var currentQuantity: Int = 0
     var filled: MutableList<OrderFiller> = mutableListOf()
     constructor(quantity: Long, type: String, price: Long, orderId: Int){
-        this.quantity = quantity
+        this.quantity = quantity.toInt()
         this.type = type
-        this.price = price
+        this.price = price.toInt()
         this.orderId = orderId
 
     }
-    fun updateOrderQuantity(quantity: Long, amount: Long){
+    fun updateOrderQuantity(quantity: Int, amount: Int){
         // This function will execute when the user's order was partially or fully
         // filled/sold
         if((currentQuantity - quantity) >= 0){
             currentQuantity -= quantity
-            if(currentQuantity == 0L){
+            if(currentQuantity == 0){
                 orderStatus = "COMPLETED"
             }
             else{
